Vaani_Sentinel/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database and ORM
sqlalchemy==2.0.35
alembic==1.13.3

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# AI and ML APIs
google-generativeai==0.3.2
groq==0.4.1
openai==1.3.7
langdetect==1.0.9
transformers==4.35.2
torch==2.5.1

# Text-to-Speech (FREE ONLY)
gTTS==2.4.0
pydub==0.25.1

# Content Processing
profanity-check==1.0.3
textblob==0.17.1
better-profanity==0.7.0

# Data Processing
pandas>=2.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.25.2
aiofiles==23.2.0
celery==5.3.4
redis==5.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# File handling
openpyxl==3.1.2
xlrd==2.0.1
